requirements.txt
streamlit>=1.31
pymupdf>=1.23
pillow>=10.0
pandas>=2.0
